comparison fix

+++ src/Cargo/src/Cargo.cpp
@@ -26,7 +26,7 @@
 
 #include <iostream>
 
-static const int ChunkSize = 1024;
+static const size_t ChunkSize = 1024;
 static const int ZipWindowBits = 15;
 static const uint16_t ZipMinVersion = 0x14;
 static const uint16_t ZipArchiveFile = 1;
